backend/requirements.txt
fastapi>=0.122.0
uvicorn==0.24.0
sqlalchemy>=2.0.44
email-validator>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
gensim>=4.3.0
joblib>=1.3.0
PyPDF2>=3.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
alembic>=1.13.0